Skip9

+++ service/todo.go
@@ -3,6 +3,8 @@ package service
 import (
 	"context"
 	"database/sql"
+	"fmt"
+	"strings"
 
 	"github.com/TechBowl-japan/go-stations/model"
 )
@@ -42,13 +44,13 @@ func (s *TODOService) CreateTODO(ctx context.Context, subject, description strin
 		return nil, err
 	}
 
-	todo := model.TODO{ID: lastId}
-	err = s.db.QueryRowContext(ctx, confirm, lastId).Scan(&todo.Subject,&todo.Description,&todo.CreatedAt,&todo.UpdatedAt)
+	todo := &model.TODO{ID: lastId}
+	err = s.db.QueryRowContext(ctx, confirm, lastId).Scan(&todo.Subject, &todo.Description, &todo.CreatedAt, &todo.UpdatedAt)
 	if err != nil {
 		return nil, err
 	}
 
-	return &todo, nil
+	return todo, nil
 }
 
 // ReadTODO reads TODOs on DB.
@@ -58,7 +60,30 @@ func (s *TODOService) ReadTODO(ctx context.Context, prevID, size int64) ([]*mode
 		readWithID = `SELECT id, subject, description, created_at, updated_at FROM todos WHERE id < ? ORDER BY id DESC LIMIT ?`
 	)
 
-	return nil, nil
+	var (
+		rows *sql.Rows
+		err  error
+	)
+	if prevID == 0 {
+		rows, err = s.db.QueryContext(ctx, read, size)
+	} else {
+		rows, err = s.db.QueryContext(ctx, readWithID, prevID, size)
+	}
+	if err != nil {
+		return nil, err
+	}
+
+	todos := []*model.TODO{}
+	for rows.Next() {
+		todo := &model.TODO{}
+		err = rows.Scan(&todo.ID, &todo.Subject, &todo.Description, &todo.CreatedAt, &todo.UpdatedAt)
+		if err != nil {
+			return nil, err
+		}
+		todos = append(todos, todo)
+	}
+
+	return todos, nil
 }
 
 // UpdateTODO updates the TODO on DB.
@@ -68,12 +93,68 @@ func (s *TODOService) UpdateTODO(ctx context.Context, id int64, subject, descrip
 		confirm = `SELECT subject, description, created_at, updated_at FROM todos WHERE id = ?`
 	)
 
-	return nil, nil
+	stmt, err := s.db.PrepareContext(ctx, update)
+	if err != nil {
+		return nil, err
+	}
+	defer stmt.Close()
+
+	result, err := stmt.Exec(subject, description, id)
+	if err != nil {
+		return nil, err
+	}
+
+	rows, err := result.RowsAffected()
+	if err != nil {
+		return nil, err
+	}
+	if rows == 0 {
+		return nil, &model.ErrNotFound{RowIDs: []int64{id}}
+	}
+
+	lastId, err := result.LastInsertId()
+	if err != nil {
+		return nil, err
+	}
+
+	todo := &model.TODO{ID: lastId}
+	err = s.db.QueryRowContext(ctx, confirm, lastId).Scan(&todo.Subject, &todo.Description, &todo.CreatedAt, &todo.UpdatedAt)
+	if err != nil {
+		return nil, err
+	}
+
+	return todo, nil
 }
 
 // DeleteTODO deletes TODOs on DB by ids.
 func (s *TODOService) DeleteTODO(ctx context.Context, ids []int64) error {
+	if len(ids) == 0 {
+		return nil
+	}
+
 	const deleteFmt = `DELETE FROM todos WHERE id IN (?%s)`
+	stmt, err := s.db.PrepareContext(ctx, fmt.Sprintf(deleteFmt, strings.Repeat(", ?", len(ids)-1)))
+	if err != nil {
+		return err
+	}
+	defer stmt.Close()
+
+	var ids_any []interface{}
+	for _, id := range ids {
+		ids_any = append(ids_any, id)
+	}
+	result, err := stmt.Exec(ids_any...)
+	if err != nil {
+		return err
+	}
+
+	rows, err := result.RowsAffected()
+	if err != nil {
+		return err
+	}
+	if rows == 0 {
+		return &model.ErrNotFound{RowIDs: ids}
+	}
 
 	return nil
 }
